requirements.txt
anthropic>=0.40.0
tavily-python>=0.5.0
apscheduler>=3.10.0
requests>=2.31.0
matplotlib>=3.8.0
numpy>=1.26.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
Pillow>=10.0.0
